Extract predicate candidate filtering
